refactor: 💡 improve uWebsocket.js HTTP integrations
