+++ src/demo/reactive-rpc/server.ts
@@ -3,8 +3,9 @@ import {enableCors, createConnectionContext, ConnectionContext, enableWsBinaryRe
 import {sampleApi} from '../../reactive-rpc/common/rpc/__tests__/api';
 import {RpcServer} from '../../reactive-rpc/common/rpc';
 import {RpcApiCaller} from '../../reactive-rpc/common/rpc/RpcApiCaller';
-import {enableHttpGetRpcApi, enableHttpPostRpcApi, UwsHttpBaseContext} from '../../reactive-rpc/server/uws/http/static';
-import { enableSsePostRpcApi } from '../../reactive-rpc/server/uws/http/sse';
+import {enableHttpGetRpcApi, enableHttpPostRpcApi} from '../../reactive-rpc/server/uws/http/static';
+import {enableSsePostRpcApi} from '../../reactive-rpc/server/uws/http/sse';
+import {UwsHttpBaseContext} from '../../reactive-rpc/server/uws/http/types';
 
 const uws = App({});
