+++ src/reactive-rpc/server/uws/http/types.ts
@@ -1,0 +1,3 @@
+export interface UwsHttpBaseContext {
+  payloadSize?: number;
+}
